requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
idna==2.10
pkg-resources==0.0.0
psycopg2==2.7.4
pyshorteners==1.0.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
stripe==2.49.0
urllib3==1.25.10